Conflicts:
  api/jsonapi_helpers_test.go
  api/jsonapi_test.go
  bitcoin/status.go
  core/core.go
  repo/db/cases_test.go
  schema/configuration.go

* master: (66 commits)
  Change default IPNS QuerySize
  BUGFIX: Fix ipfs keystore caching.
  Bump minor version number
  TWEAK: Add ob/inventory to gateway allowed GETs.
  TWEAK: Include shippingOption with digital and service orders for legacy compatibility.
  Add selfhostedstorage_test
  Update inventory_test.go
  Create GetOrderQuantity function
  Modify InventoryStore interface to use an int64 rather than an int
  Add uint64 quantity field to order
  Fix test failure string in OutOfInventoryTest
  Fix bug in multiaddr dependency
  TESTS: Update qa test to work with new inventory format.
  BUGFIX: Update shipping handling to support newer listing versions.
  TWEAK: Use helper func for API errors.
  FEATURE: Make out of inventory errors JSON messages that include remaining inventory.
  TESTS: Rename alice/bob to vendor/buyer in purchase_direct_online qa test.
  Revert "Revert "REFACTOR: Update local inventory API response to mirror IPFS inventory response.""
  TWEAK: Bump listing version number.
  Revert "REFACTOR: Update local inventory API response to mirror IPFS inventory response."
  ...

+++ repo/datastore.go
@@ -160,16 +160,16 @@ type InventoryStore interface {
 
 	/* Put an inventory count for a listing
 	   Override the existing count if it exists */
-	Put(slug string, variantIndex int, count int) error
+	Put(slug string, variantIndex int, count int64) error
 
 	// Return the count for a specific listing including variants
-	GetSpecific(slug string, variantIndex int) (int, error)
+	GetSpecific(slug string, variantIndex int) (int64, error)
 
 	// Get the count for all variants of a given listing
-	Get(slug string) (map[int]int, error)
+	Get(slug string) (map[int]int64, error)
 
 	// Fetch all inventory maps for each slug
-	GetAll() (map[string]map[int]int, error)
+	GetAll() (map[string]map[int]int64, error)
 
 	// Delete a listing and related count
 	Delete(slug string, variant int) error
